+++ pages/_app.js
@@ -1,15 +1,4 @@
 import '../styles/globals.css'
-// import "@fortawesome/fontawesome-svg-core/styles.css"; // import Font Awesome CSS
-// import { config } from "@fortawesome/fontawesome-svg-core";
-// config.autoAddCss = false; // Tell Font Awesome to skip adding the CSS automatically since it's being imported above
-
-import ReactDOM from 'react-dom';
-import { library } from '@fortawesome/fontawesome-svg-core'; //fontawesomeのコアファイル
-import { fab } from '@fortawesome/free-brands-svg-icons'; //fontawesomeのbrandアイコンのインポート
-import { fas } from '@fortawesome/free-solid-svg-icons'; //fontawesomeのsolidアイコンのインポート
-import { far } from '@fortawesome/free-regular-svg-icons'; //fontawesomeのregularアイコンのインポート
-library.add(fab, fas, far); //他のコンポーネントから簡単に呼び出せるようにするための登録処理？
-import '@fontsource/roboto';
 
 function MyApp({ Component, pageProps }) {
